Greybus: Hook the bundle devices into linux suspend framework.
This patch hooks the bundle devices to the linux suspend framework.
The bundle suspend API in turn calls into the suspend APIs registered
by the underlying connections before attempting to power off the interface.
It is the responsibility of the connections in a bundle to register
connection suspend APIs and tranistion the connection to a proper
power state.

ToDo: Note that resume is not implemented as the suspend will eventually
cut the power to the interface and all the devices will disappear. A resume
will be like a cold boot and not through corresponding bundle resume APIs.
One way of achieving this is by implementing a resume hook for the host
device which can maintain a list of previously attached interfaces and
proceed to turn the power and reference clock to the interfaces. This should
take the interfaces through the cold boot sequence. This is a suggestion and
is not implemented today.

Signed-off-by: Thara Gopinath <thara.gopinath@linaro.org>