Greybus: Sample implementation of linux suspend in I2C connection.
This patch provides a sample implementation of how connection drivers have to
define the suspend hook that can be called from the bundle linux suspend
API and the mandatory steps the connection suspend hook has to implement.

This is not a complete implementation of suspend API as the driver
may have to do additional steps before giving the go ahead to power off
the connection and subsequently the bundle and the interface.

Signed-off-by: Thara Gopinath <thara.gopinath@linaro.org>